Dynamic render of products in SPA implementation
