+++ index.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="css/style.css">
     <script src="https://kit.fontawesome.com/577c267cd5.js" crossorigin="anonymous" defer></script>
-    <script  type="module" src="dist/bundle.js" defer></script>
+    <script  type="module" src="dist/bundle.js"></script>
     <!--<script type="module" src="js/app.js" defer></script>-->
     <title>Gfx Store</title>
 </head>
